clean package
