+++ app/src/main/java/org/positive/daymotion/presentation/home/HomeTabViewModel.kt
@@ -1,8 +1,8 @@
-package org.positive.daymotion.presentation.root.tabs.my
+package org.positive.daymotion.presentation.home
 
 import dagger.hilt.android.lifecycle.HiltViewModel
 import org.positive.daymotion.presentation.base.BaseViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class MyTabViewModel @Inject constructor() : BaseViewModel()
+class HomeTabViewModel @Inject constructor() : BaseViewModel()
