+++ app/src/main/java/org/positive/daymotion/presentation/my/MyTabFragment.kt
@@ -1,17 +1,17 @@
-package org.positive.daymotion.presentation.root.tabs.setting
+package org.positive.daymotion.presentation.my
 
 import android.os.Bundle
 import android.view.View
 import dagger.hilt.android.AndroidEntryPoint
 import org.positive.daymotion.R
-import org.positive.daymotion.databinding.FragmentSettingTabBinding
+import org.positive.daymotion.databinding.FragmentMyTabBinding
 import org.positive.daymotion.presentation.base.BaseFragment
 import org.positive.daymotion.presentation.base.util.viewModelOf
 
 @AndroidEntryPoint
-class SettingTabFragment : BaseFragment<FragmentSettingTabBinding>(R.layout.fragment_setting_tab) {
+class MyTabFragment : BaseFragment<FragmentMyTabBinding>(R.layout.fragment_my_tab) {
 
-    private val viewModel by viewModelOf<SettingTabViewModel>()
+    private val viewModel by viewModelOf<MyTabViewModel>()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
