+++ app/src/main/java/org/positive/daymotion/presentation/my/MyTabViewModel.kt
@@ -1,8 +1,8 @@
-package org.positive.daymotion.presentation.root.tabs.category
+package org.positive.daymotion.presentation.my
 
 import dagger.hilt.android.lifecycle.HiltViewModel
 import org.positive.daymotion.presentation.base.BaseViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class CategoryTabViewModel @Inject constructor() : BaseViewModel()
+class MyTabViewModel @Inject constructor() : BaseViewModel()
