+++ app/src/main/java/org/positive/daymotion/presentation/setting/SettingTabViewModel.kt
@@ -1,8 +1,8 @@
-package org.positive.daymotion.presentation.root.tabs.home
+package org.positive.daymotion.presentation.setting
 
 import dagger.hilt.android.lifecycle.HiltViewModel
 import org.positive.daymotion.presentation.base.BaseViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class HomeTabViewModel @Inject constructor() : BaseViewModel()
+class SettingTabViewModel @Inject constructor() : BaseViewModel()
